% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataCSV}
\alias{exampleDataCSV}
\title{Example CSV data description.}
\format{
.csv
}
\source{
exampleRepo
}
\usage{
exampleDataCSV
}
\description{
Example CSV data description.
}
\examples{
\dontrun{
 library(exampleRepo2);
 exampleRepo2::exampleDataCSV
}
}
\keyword{datasets}
