% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataR}
\alias{exampleDataR}
\title{Example R data description.}
\format{
R
}
\source{
exampleRepo
}
\usage{
exampleDataR
}
\description{
Example R data description.
}
\examples{
\dontrun{
 library(exampleRepo2);
 exampleRepo2::exampleDataR
}
}
\keyword{datasets}
