% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRepo2.R
\docType{package}
\name{exampleRepo2}
\alias{exampleRepo2}
\title{exampleRepo2}
\description{
This is a package description.
}
\author{
Zarrar Khan
\itemize{
\item Github: https://github.com/zarrarkhan/exampleRepo2
\item Webpage: https://zarrarkhan.github.io/exampleRepo2/}
}
